MDEV-20595: Assertion `0 < sel && sel <= 2.0' failed in table_cond_selectivity
Selectivity is discounted twice for a field when the field has a condition like t1.col1=t2.col1
and t2.col1=const.
So if there is ref access on table t1, it should be of type ref(const) [equality propagation is done already]
But in the function table_cond_selectivity where we discount the selectivity for ref(const),
we also try to discount the selectivity for the condition t1.col=t2.col1, which is incorrect.
Just discount the selectivity for a field only ONCE.

Fixed by using the tmp_set bitmap that would hold the indexes of fields for which selectivity
was already discounted.